Refactor documentation and code structure for security features and integrations
- Removed outdated design specifications for Wave 3 security features.
- Updated integrations documentation to clarify Azure AD usage.
- Revised Kubernetes README to point to the correct deployment documentation.
- Enhanced mkdocs configuration for improved navigation and structure.
- Consolidated documentation for OAuth2 server, including endpoint updates and security features.
- Merged tracing and logging documentation into a single file for clarity.
- Slimmed down README and index pages to reduce redundancy and improve readability.
- Expanded admin stub pages with detailed content for better guidance.

+++ crates/oauth2-actix/src/actors/client_actor.rs
@@ -208,7 +208,7 @@ impl Handler<GetClient> for ClientActor {
         #[cfg(feature = "redis-cache")]
         let redis_conn = self.redis.clone();
         #[cfg(feature = "redis-cache")]
-        let cache_ttl_secs = self.cache_ttl.as_secs().max(1) as u64;
+        let cache_ttl_secs = self.cache_ttl.as_secs().max(1);
 
         Box::pin(
             async move {
@@ -303,7 +303,7 @@ impl Handler<ValidateClient> for ClientActor {
         #[cfg(feature = "redis-cache")]
         let redis_conn = self.redis.clone();
         #[cfg(feature = "redis-cache")]
-        let cache_ttl_secs = self.cache_ttl.as_secs().max(1) as u64;
+        let cache_ttl_secs = self.cache_ttl.as_secs().max(1);
 
         Box::pin(
             async move {
